# api/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "3.0.0",
        "title": "Books API",
        "description": "THis api is for books",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:5000",
    "basePath": "/v1/db",
    "tags": [
        {
            "name": "books",
            "description": "Everything about Books"
        }
    ],
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get all books",
                "description": "Returns all books",
                "operationId": "getBooks",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of the books",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset of the books",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Book"
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get book by id",
                "description": "Returns a book",
                "operationId": "getBookById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/BookDetails"
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get author by id",
                "description": "Returns an author",
                "operationId": "getAuthorById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the author",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/subjects/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get subject by id",
                "description": "Returns a subject",
                "operationId": "getSubjectById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the subject",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Search books",
                "description": "Returns a book",
                "operationId": "searchBooks",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of the books",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset of the books",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Book"
                            }
                        }
                    }
                }
            }
        },
        "/subjects": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get all subjects",
                "description": "Returns all subjects",
                "operationId": "getSubjects",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of the subjects",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset of the subjects",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/bysubject": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get books by subject",
                "description": "Returns all books by subject",
                "operationId": "getBooksBySubject",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "Subject of the books",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of the books",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset of the books",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Book"
                            }
                        }
                    }
                }
            }
        },
        "/gettext/{id}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get text of the book",
                "description": "Returns text of the book",
                "operationId": "getText",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the book",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Book": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "title": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer",
                    "format": "int64"
                },
                "cover": {
                    "type": "string"
                }
            }
        },
        "BookDetails": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "title": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer",
                    "format": "int64"
                },
                "cover": {
                    "type": "string"
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "copyright": {
                    "type": "string"
                }
            }
        }
    }
}